#45 Added the FDPS calls to the HDF5 IO module
